backend/requirements.txt
# Flask Core - FREE
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database - PostgreSQL (FREE, open-source)
# Database
psycopg2-binary>=2.9.9

# LLM - Google Gemini (FREE tier: 1500 req/day)
google-generativeai==0.3.2

# LLM - Ollama (100% FREE, optional for local LLM)
# Uncomment if you want to use completely free local LLM
# ollama==0.1.6

# Security
werkzeug==3.0.1
bcrypt==4.1.2
Flask-Limiter==3.5.0

# Utilities
requests==2.31.0

# Development (optional)
pytest==7.4.3
pytest-flask==1.3.0
